# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for simpo pharmacy management system supporting POS, inventory, reporting, and multi-branch operations.",
        "title": "simpo Pharmacy Management System API",
        "termsOfService": "https://simpo.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@simpo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/admin/audit/retention/cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes audit logs older than 5 years with optional backup (Story 6.4, AC6)\nCRITICAL: This operation permanently deletes audit logs. Use with caution.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Log Management"
                ],
                "summary": "Execute audit log retention cleanup",
                "parameters": [
                    {
                        "description": "Cleanup request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CleanupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cleanup executed successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.CleanupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or not confirmed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/admin/audit/retention/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns information about audit logs eligible for cleanup (Story 6.4, AC6)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit Log Management"
                ],
                "summary": "Get audit log retention status",
                "responses": {
                    "200": {
                        "description": "Retention status retrieved",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetRetentionStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - SYSTEM_ADMIN only",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CleanupRequest": {
            "type": "object",
            "required": [
                "confirmed",
                "reason"
            ],
            "properties": {
                "confirmed": {
                    "description": "Confirmed must be true to execute the cleanup (safety confirmation)\nExample: true",
                    "type": "boolean",
                    "example": true
                },
                "create_backup": {
                    "description": "CreateBackup indicates whether to create a backup before cleanup\nExample: true",
                    "type": "boolean",
                    "example": true
                },
                "reason": {
                    "description": "Reason is the reason for performing cleanup (minimum 10 characters, maximum 500)\nExample: \"Scheduled quarterly cleanup per compliance policy\"\nCode review fix: CRIT-015 - Add max length validation",
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 10,
                    "example": "Scheduled quarterly cleanup per compliance policy"
                }
            }
        },
        "handlers.CleanupResponse": {
            "type": "object",
            "properties": {
                "backup_file": {
                    "description": "BackupFile is the backup file created (if CreateBackup was true)\nExample: \"audit_backup_20260527.csv\"",
                    "type": "string",
                    "example": "audit_backup_20260527.csv"
                },
                "deleted_count": {
                    "description": "DeletedCount is the number of audit log entries deleted\nExample: 5000",
                    "type": "integer",
                    "example": 5000
                },
                "executed_at": {
                    "description": "ExecutedAt is the timestamp when cleanup was executed\nExample: \"2026-05-27T15:00:00Z\"",
                    "type": "string",
                    "example": "2026-05-27T15:00:00Z"
                },
                "reason": {
                    "description": "Reason is the reason provided for the cleanup\nExample: \"Scheduled quarterly cleanup per compliance policy\"",
                    "type": "string",
                    "example": "Scheduled quarterly cleanup per compliance policy"
                }
            }
        },
...
        "handlers.GetRetentionStatusResponse": {
            "type": "object",
            "properties": {
                "cutoff_date": {
                    "description": "CutoffDate is the date before which logs are eligible for cleanup\nExample: \"2021-05-27T00:00:00Z\"",
                    "type": "string",
                    "example": "2021-05-27T00:00:00Z"
                },
                "eligible_for_cleanup": {
                    "description": "EligibleForCleanup is the number of logs older than 5 years",
                    "type": "integer",
                    "example": 5000
                },
                "estimated_space_reclaimed": {
                    "description": "EstimatedSpaceReclaimed is the estimated disk space that would be reclaimed (in bytes)\nExample: 52428800",
                    "type": "integer",
                    "example": 52428800
                },
                "total_logs": {
                    "description": "TotalLogs is the total number of audit logs in the system",
                    "type": "integer",
                    "example": 150000
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}